+++ core/src/types/transport_params.rs
@@ -11,7 +11,7 @@ pub struct CreateGameAccountParams {
     pub title: String,
     pub bundle_addr: String,
     pub token_addr: String,
-    pub max_players: u8,
+    pub max_players: u16,
     pub min_deposit: u64,
     pub max_deposit: u64,
     pub data: Vec<u8>,
